requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0.35
greenlet>=3.0.3
psycopg2-binary
python-dotenv
email-validator
requests
httpx
celery
redis
argon2-cffi
PyJWT
prometheus-client
langchain
langchain-openai
langchain-google-genai
google-cloud-vision
langfuse
litellm
qdrant-client
pypdf
PyMuPDF # PDF 구조 분석 및 표 추출을 위해 추가
xmltodict
watchdog
beautifulsoup4
minio
python-multipart
lxml
feedparser
lxml
PyYAML
google-auth==2.33.0
pyotp==2.9.0
